+++ application/shared/view/timeclock/employee/view.php
@@ -1,7 +1,3 @@
-<?php
-$date_format = 'm/d/y';
-?>
-
 <div class="jumbotron">
     <h1>TimeClock</h1>
     <p>View employee</p>