+++ application/shared/view/timeclock/jobs/view.php
@@ -12,14 +12,25 @@ $date_format = 'm/d/y';
             <div class="panel panel-primary">
                 <div class="panel-heading">
                     <h3 class="panel-title row">
-                        <div class="title col-sm-5">
+                        <div class="title col-sm-6">
                             {job['job_name']} ({job['job_uid']}) for {job['client_name']}
-                            (<span class="start_date">{start_date}</span> -
-                            <span class="end_date">{last_date}</span>) -
-                            <?php //echo (float) $this->sys->template->total_hours; ?> hours
+                            <?php
+                            switch ($this->sys->template->job['status']) {
+                                case 'c':
+                                    $status = 'Completed';
+                                    break;
+                                case 'wip':
+                                    $status = "In Progress";
+                                    break;
+                                default: //na
+                                    $status = "Not Started";
+                            }
+                            ?>
+                            (<?php echo $status; ?> <span class="start_date">{start_date}</span> -
+                            <span class="end_date">{last_date}</span>)
                         </div>
-                        <div class="col-sm-1 col-sm-offset-6">
-                            <a href="{timeclock_root}jobs/print_friendly/{jobs['job_id']}" target="_blank" class="btn btn-primary btn-sm" role="button">Print</a>
+                        <div class="col-sm-1 col-sm-offset-5">
+                            <a href="{timeclock_root}jobs/print_friendly/{job['job_id']}" target="_blank" class="btn btn-primary btn-sm" role="button">Print</a>
                         </div>
                     </h3>
                 </div>
@@ -50,8 +61,47 @@ $date_format = 'm/d/y';
                         </tfoot>
                     </table>
                     <a href="javascript:addDate({job_id}, 'job')">Add Date</a>
-                </div>
-            </div> <!-- END: panel-body -->
+                </div> <!-- END: panel-body -->
+            </div> <!-- END: panel -->
+            <div class="row">
+                <div class="col-sm-6">
+                    <div class="panel panel-success">
+                        <div class="panel-heading">
+                            <h3 class="panel-title row">
+                                <div class="title col-sm-12">
+                                    Hours Breakdown ({total_hours} Total Hours)
+                                </div>
+                            </h3>
+                        </div>
+                        <div class="panel-body">
+                            <table class="table table-striped">
+                                <thead>
+                                    <tr>
+                                        <th>Category</th>
+                                        <th>Total Hours</th>
+                                    </tr>
+                                </thead>
+                                <tbody>
+                                    <?php
+                                    foreach ($this->sys->template->hours_by_category as $category=>$time) {
+                                        echo '<tr>';
+                                        echo '<td>' . $category . '</td>';
+                                        echo '<td>' . $time . '</td>';
+                                        echo '</tr>';
+                                    }
+                                    ?>
+                                </tbody>
+                                <tfoot>
+                                    <tr>
+                                        <th>Category</th>
+                                        <th>Total Hours</th>
+                                    </tr>
+                                </tfoot>
+                            </table>
+                        </div> <!-- END: col-sm -->
+                    </div> <!-- END: row -->
+                </div> <!-- END: panel-body -->
+            </div> <!-- END: panel -->
         </div> <!-- END: well -->
     </div> <!-- END: col-sm-12 -->
 </div> <!-- END: row -->